bitonic: use empty-struct chan
This CL uses 'chan struct{}' as the type for the sentinel channel.
This is more Go-idiomatic for a sentinel channel (where only the fact that
something transited thru the channel matters, not its value) and, as a value
of type struct{} does not use any memory, it should reduce the memory footprint.